Add support multiple config for token providers (#99)
* Add support multiple config for token providers

- This is a change to the existing functionality but is backwards compatible and hence feature parity.
- Code before this change was supporting just one configuration for an ID provider. So, if an org had multiple azure configurations which needed supporting, that was not possible. The usecase could be to have different azure configuration for say prod and non-prod.
- The configurations could now be provided as a list against the ID provider. The targets can then be added within that configuration rather than a global configuration for all the targets which was the case before. Readme has been updated to reflect the new configuration along with the deprecation information of the old config format.

+++ e2e/e2e_suite_test.go
@@ -49,16 +49,17 @@ var (
 	testDir        string
 
 	// Suite variables modifiable per test scenario
-	err                error
-	environmentsToUse  map[string][]string
-	targetedOspreys    []*ospreytest.TestOsprey
-	ospreyconfig       *ospreytest.TestConfig
-	ospreyconfigFlag   string
-	defaultGroup       string
-	targetGroup        string
-	targetGroupFlag    string
-	apiServerURL       string
-	useGKEClientConfig bool
+	err                    error
+	environmentsToUse      map[string][]string
+	targetedOspreys        []*ospreytest.TestOsprey
+	ospreyconfig           *ospreytest.TestConfig
+	ospreyconfigFlag       string
+	legacyOspreyconfigFlag string
+	defaultGroup           string
+	targetGroup            string
+	targetGroupFlag        string
+	apiServerURL           string
+	useGKEClientConfig     bool
 )
 
 var _ = BeforeSuite(func() {
@@ -112,6 +113,7 @@ func setupClientForEnvironments(providerName string, envs map[string][]string, c
 	ospreyconfig, err = ospreytest.BuildConfig(testDir, providerName, defaultGroup, envs, ospreys, clientID, apiServerURL, useGKEClientConfig)
 	Expect(err).To(BeNil(), "Creates the osprey config with groups")
 	ospreyconfigFlag = "--ospreyconfig=" + ospreyconfig.ConfigFile
+	legacyOspreyconfigFlag = "--ospreyconfig=" + ospreyconfig.LegacyConfigFile
 
 	if targetGroup != "" {
 		targetGroupFlag = "--group=" + targetGroup
@@ -133,7 +135,12 @@ func resetDefaults() {
 }
 
 func cleanup() {
-	if err := os.Remove(ospreyconfig.Kubeconfig); err != nil {
-		Expect(os.IsNotExist(err)).To(BeTrue())
+	if ospreyconfig != nil {
+		if err := os.Remove(ospreyconfig.Kubeconfig); err != nil {
+			Expect(os.IsNotExist(err)).To(BeTrue())
+		}
+		if err := os.Remove(ospreyconfig.LegacyConfig.Kubeconfig); err != nil {
+			Expect(os.IsNotExist(err)).To(BeTrue())
+		}
 	}
 }
